# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/caves": {
            "get": {
                "tags": [
                    "Cave"
                ],
                "summary": "getCaveList",
                "description": "Get all Caves",
                "operationId": "14757fb12ad2e8397562b4882969690f",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cave"
                ],
                "summary": "createCave",
                "description": "Create Cave",
                "operationId": "577a84bf0de4130a9b10d13f0c622bf1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/caves/{id}": {
            "get": {
                "tags": [
                    "Cave"
                ],
                "summary": "getCaveItem",
                "description": "Get Cave",
                "operationId": "1c622d9db9c463a33aa532b32f3d23de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Cave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cave"
                ],
                "summary": "updateCave",
                "description": "Update Cave",
                "operationId": "704e65587b6fff0267744aa4bc2adaba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Cave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cave"
                ],
                "summary": "deleteCave",
                "description": "Delete Cave",
                "operationId": "ea8453b16a586fecc4057c830d319683",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Cave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}